deps/v8/test/inspector/runtime/preview-sideeffects-expected.txt

Checks that console preview does not run user code

Running test: ObjectGetter
{
    getterCalled : false
}
